+++ docs/changelog.md
@@ -17,6 +17,7 @@
 - Added hardware file `hardware.ni_x_series.ni_x_series_counter` to use NI 63xx cards as fastcounters for pulsed measurements.
 - Added resizeable `PulseBlockEditor` columns in `pulsed_gui`
 - Added pulse shaping to the `pulsed` tool chain
+- supporting custom alternative plots in pulsed analysis
 
 ### Other
 - Displaying Qudi logs in the test workflow upon test failure